Copy tree ID from gold tree when creating result
The parser creates the output trees in the same order as the given gold trees, but gives them IDs according to the order and ignoring the original IDs given in the gold trees.
Here I change the behavior to copy the gold tree IDs, to ease evaluation by external tools (which will now simply be able to take the trees with the same ID and will not have to count them).